# docs/openapi/groupshop-api.yaml
openapi: 3.0.3
info:
  title: GroupShop MVP API
  version: 0.10.0
  description: Machine-checkable API contract for implemented MVP batches.
servers:
  - url: /
...
paths:
...
  /api/v1/cart/items:
    get:
      tags:
        - Cart
      summary: List cart items.
      operationId: listCartItems
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart items listed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItemsSuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Cart
      summary: Add item to cart.
      operationId: addCartItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemRequest'
      responses:
        '200':
          description: Item added to cart.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItemSuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          description: Business rule violation (e.g. GROUP_BUY_ENDED, INSUFFICIENT_STOCK).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/cart/items/{cartItemId}:
    patch:
      tags:
        - Cart
      summary: Update cart item quantity.
      operationId: updateCartItem
      security:
        - bearerAuth: []
      parameters:
        - name: cartItemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
      responses:
        '200':
          description: Cart item updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItemSuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/CartForbiddenError'
        '404':
          $ref: '#/components/responses/CartNotFoundError'
        '422':
          description: Business rule violation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Cart
      summary: Delete a cart item.
      operationId: deleteCartItem
      security:
        - bearerAuth: []
      parameters:
        - name: cartItemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cart item deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptySuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/CartForbiddenError'
        '404':
          $ref: '#/components/responses/CartNotFoundError'
  /api/v1/cart/items/delete-all:
    delete:
      tags:
        - Cart
      summary: Clear all cart items for the current user.
      operationId: clearCart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart cleared.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptySuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/cart/checkout-preview:
    post:
      tags:
        - Cart
      summary: Preview order from cart items.
      operationId: cartCheckoutPreview
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartCheckoutPreviewRequest'
      responses:
        '200':
          description: Checkout preview data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPreviewSuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
...
  /api/v1/my/orders/{orderId}:
    get:
      tags:
        - Order
      summary: Get order detail.
      operationId: getMyOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            description: ID（string）
      responses:
        '200':
          description: Order detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/v1/my/store/orders:
    get:
      tags:
        - StoreOrder
      summary: List orders for the current user's store.
      operationId: listStoreOrders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated store order list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPageSuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/LeaderRequiredError'
  /api/v1/my/store/orders/{orderId}:
    get:
      tags:
        - StoreOrder
      summary: Get store order detail.
      operationId: getStoreOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            description: ID（string）
      responses:
        '200':
          description: Store order detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/LeaderRequiredError'
          # Also: STORE_FORBIDDEN
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/my/store/orders/{orderId}/ship:
    post:
      tags:
        - StoreOrder
      summary: Ship an order (create shipment, mark order as shipped).
      operationId: shipOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            description: ID（string）
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
          description: Optional idempotency key for safe retry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShipOrderRequest'
      responses:
        '200':
          description: Order shipped successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipOrderSuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/LeaderRequiredError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/OrderAlreadyShippedError'
        '422':
          $ref: '#/components/responses/OrderNotShippableError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    LeaderRequiredError:
      description: The current operation requires a leader/store identity.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BusinessRuleViolation:
      description: The operation violates a business rule.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    CartForbiddenError:
      description: The cart item does not belong to the current user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    CartNotFoundError:
      description: The cart item was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    OrderAlreadyShippedError:
      description: Order has already been shipped.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    OrderNotShippableError:
      description: Order is not in a shippable state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    EmptySuccessResponse:
      type: object
      required:
        - success
        - traceId
      properties:
        success:
          type: boolean
          enum:
            - true
        traceId:
          type: string
          minLength: 1
      additionalProperties: false
    # ── Error ─────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - traceId
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          $ref: '#/components/schemas/ErrorDetail'
        traceId:
          type: string
          minLength: 1
      additionalProperties: false
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        details:
          type: object
          nullable: true
      additionalProperties: false
    # ── Health ────────────────────────────────────────────────────
...
    OrderPreviewSuccessResponse:
      type: object
      required:
        - success
        - data
        - traceId
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/OrderPreviewData'
        traceId:
          type: string
      additionalProperties: false
    OrderPreviewData:
      type: object
      required:
        - groupBuyId
        - address
        - items
        - totalAmount
        - discountAmount
        - payAmount
      properties:
        groupBuyId:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        address:
          $ref: '#/components/schemas/OrderPreviewAddress'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderPreviewItem'
        totalAmount:
          type: integer
          minimum: 0
        discountAmount:
          type: integer
          minimum: 0
        payAmount:
          type: integer
          minimum: 0
      additionalProperties: false
    OrderPreviewAddress:
      type: object
      required:
        - id
        - receiverName
        - receiverPhone
        - province
        - city
        - district
        - detail
        - fullAddress
      properties:
        id:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        receiverName:
          type: string
        receiverPhone:
          type: string
        province:
          type: string
        city:
          type: string
        district:
          type: string
        detail:
          type: string
        fullAddress:
          type: string
      additionalProperties: false
    OrderPreviewItem:
      type: object
      required:
        - groupBuyItemId
        - productId
        - productName
        - unitPriceAmount
        - quantity
        - totalAmount
        - availableStock
        - soldCount
      properties:
        groupBuyItemId:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        productId:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        productName:
          type: string
        skuName:
          type: string
        unitPriceAmount:
          type: integer
          minimum: 0
        quantity:
          type: integer
          minimum: 1
        totalAmount:
          type: integer
          minimum: 0
        availableStock:
          type: integer
          minimum: 0
        soldCount:
          type: integer
          minimum: 0
      additionalProperties: false
...
    OrderSuccessResponse:
      type: object
      required:
        - success
        - data
        - traceId
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/OrderData'
        traceId:
          type: string
      additionalProperties: false
    OrderPageSuccessResponse:
      type: object
      required:
        - success
        - data
        - traceId
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/OrderPageData'
        traceId:
          type: string
      additionalProperties: false
    OrderPageData:
      type: object
      required:
        - items
        - page
        - pageSize
        - total
        - hasMore
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderData'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        hasMore:
          type: boolean
      additionalProperties: false
    OrderData:
      type: object
      required:
        - id
        - orderNo
        - groupBuyId
        - storeId
        - leaderId
        - totalAmount
        - discountAmount
        - payAmount
        - payStatus
        - orderStatus
      properties:
        id:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        orderNo:
          type: string
        groupBuyId:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        storeId:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        leaderId:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        totalAmount:
          type: integer
          minimum: 0
        discountAmount:
          type: integer
          minimum: 0
        payAmount:
          type: integer
          minimum: 0
        payStatus:
          type: string
        orderStatus:
          type: string
        paidAt:
          type: string
          nullable: true
        shippedAt:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        remark:
          type: string
          nullable: true
        receiverName:
          type: string
        receiverPhone:
          type: string
        province:
          type: string
        city:
          type: string
        district:
          type: string
        detail:
          type: string
        fullAddress:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemData'
      additionalProperties: false
    OrderItemData:
      type: object
      required:
        - id
        - groupBuyItemId
        - productId
        - productName
        - unitPriceAmount
        - quantity
        - totalAmount
      properties:
        id:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        groupBuyItemId:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        productId:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        productName:
          type: string
        skuName:
          type: string
        unitPriceAmount:
          type: integer
          minimum: 0
        quantity:
          type: integer
          minimum: 1
        totalAmount:
          type: integer
          minimum: 0
      additionalProperties: false
...
    ShipOrderRequest:
      type: object
      required:
        - deliveryType
      properties:
        deliveryType:
          type: string
          enum:
            - express
            - pickup
            - local_delivery
        logisticsCompany:
          type: string
          nullable: true
        trackingNo:
          type: string
          nullable: true
      additionalProperties: false
    ShipOrderSuccessResponse:
      type: object
      required:
        - success
        - data
        - traceId
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/ShipOrderData'
        traceId:
          type: string
      additionalProperties: false
    ShipOrderData:
      type: object
      required:
        - order
        - shipment
      properties:
        order:
          $ref: '#/components/schemas/OrderData'
        shipment:
          $ref: '#/components/schemas/ShipmentData'
      additionalProperties: false
    ShipmentData:
      type: object
      required:
        - id
        - orderId
        - deliveryType
      properties:
        id:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        orderId:
          type: string
          description: 雪花ID（字符串防JS精度丢失）
        deliveryType:
          type: string
        logisticsCompany:
          type: string
          nullable: true
        trackingNo:
          type: string
          nullable: true
        shippedAt:
          type: string
          nullable: true
      additionalProperties: false
    # ── Subscription ───────────────────────────────────────────────────
...
    CartItemData:
      type: object
      properties:
        cartItemId:
          type: string
        groupBuyId:
          type: string
        groupBuyItemId:
          type: string
        productId:
          type: string
        title:
          type: string
        coverImageUrl:
          type: string
          nullable: true
        groupPriceAmount:
          type: integer
        quantity:
          type: integer
        availableStock:
          type: integer
        visibility:
          type: string
        status:
          type: string
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
      additionalProperties: false
    AddCartItemRequest:
      type: object
      required:
        - groupBuyItemId
        - quantity
      properties:
        groupBuyItemId:
          type: string
        quantity:
          type: integer
          minimum: 1
        shareToken:
          type: string
      additionalProperties: false
    UpdateCartItemRequest:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer
          minimum: 1
        shareToken:
          type: string
      additionalProperties: false
    CartCheckoutPreviewRequest:
      type: object
      required:
        - addressId
        - cartItemIds
      properties:
        addressId:
          type: string
        cartItemIds:
          type: array
          items:
            type: string
        shareToken:
          type: string
      additionalProperties: false
    CartItemsSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/CartItemData'
      additionalProperties: false
    CartItemSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/CartItemData'
      additionalProperties: false